+++ app/services/route_type_service.py
@@ -1,5 +1,6 @@
 from app.models import RouteType, RouteTypeSchema
 from app.extensions import db
+from flask import abort
 
 route_type_schema = RouteTypeSchema()
 route_types_schema = RouteTypeSchema(many=True)
@@ -17,14 +18,18 @@ def get_route_type_by_id(route_type_id):
     return None
 
 
-def create_route_type(body):
+def create_route_type(token_info, body):
+    if token_info['role'] != 'admin':
+        abort(401, 'Unauthorized')
     new_route_type = RouteType(body['route_type_name'])
     db.session.add(new_route_type)
     db.session.commit()
     return route_type_schema.dump(new_route_type)
 
 
-def update_route_type(route_type_id, body):
+def update_route_type(token_info, route_type_id, body):
+    if token_info['role'] != 'admin':
+        abort(401, 'Unauthorized')
     route_type = RouteType.query.get(route_type_id)
     if route_type:
         route_type.route_type_name = body['route_type_name']
@@ -33,7 +38,9 @@ def update_route_type(route_type_id, body):
     return None
 
 
-def delete_route_type(route_type_id):
+def delete_route_type(token_info, route_type_id):
+    if token_info['role'] != 'admin':
+        abort(401, 'Unauthorized')
     route_type = RouteType.query.get(route_type_id)
     if route_type:
         db.session.delete(route_type)
