+++ app/services/tag_service.py
@@ -1,5 +1,6 @@
 from app.models import Tag, TagSchema
 from app.extensions import db
+from flask import abort
 
 tag_schema = TagSchema()
 tags_schema = TagSchema(many=True)
@@ -17,14 +18,18 @@ def get_tag_by_id(tag_id):
     return None
 
 
-def create_tag(body):
+def create_tag(token_info, body):
+    if token_info['role'] != 'admin':
+        abort(401, 'Unauthorized')
     new_tag = Tag(body['tag_name'])
     db.session.add(new_tag)
     db.session.commit()
     return tag_schema.dump(new_tag)
 
 
-def update_tag(tag_id, body):
+def update_tag(token_info, tag_id, body):
+    if token_info['role'] != 'admin':
+        abort(401, 'Unauthorized')
     tag = Tag.query.get(tag_id)
     if tag:
         tag.tag_name = body['tag_name']
@@ -33,7 +38,9 @@ def update_tag(tag_id, body):
     return None
 
 
-def delete_tag(tag_id):
+def delete_tag(token_info, tag_id):
+    if token_info['role'] != 'admin':
+        abort(401, 'Unauthorized')
     tag = Tag.query.get(tag_id)
     if tag:
         db.session.delete(tag)
